Updated for use of paptools project.

+++ include/fibonacci/fibonacci.h
@@ -7,22 +7,22 @@ namespace fibonacci {
 // Returns the Nth Fibonacci number using a naive recursive implementation.
 // Time: O(2^n)
 // Space: O(n)
-uint64_t RecursiveNaive(uint8_t n);
+unsigned long long RecursiveNaive(unsigned short n);
 
 // Returns the Nth Fibonacci number using a memoized recursive implementation.
 // Time: O(n)
 // Space: O(n)
-uint64_t RecursiveMemo(uint8_t n);
+unsigned long long RecursiveMemo(unsigned short n);
 
 // Returns the Nth Fibonacci number using an iterative implementation.
 // Time: O(n)
 // Space: O(1)
-uint64_t Iterative(uint8_t n);
+unsigned long long Iterative(unsigned short n);
 
 // Returns the Nth Fibonacci number using a lookup table implementation.
 // Time: O(1)
 // Space: O(1)
-uint64_t LookupTable(uint8_t n);
+unsigned long long LookupTable(unsigned short n);
 } // namespace fibonacci
 
 #endif // FIBONACCI_FIBONACCI_H
